+++ paymentsheet/src/main/java/com/stripe/android/customersheet/analytics/DefaultCustomerSheetEventReporter.kt
@@ -4,6 +4,7 @@ import com.stripe.android.core.injection.IOContext
 import com.stripe.android.core.networking.AnalyticsRequestExecutor
 import com.stripe.android.core.networking.AnalyticsRequestFactory
 import com.stripe.android.customersheet.CustomerSheet
+import com.stripe.android.customersheet.CustomerSheetIntegration
 import com.stripe.android.model.CardBrand
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
@@ -15,9 +16,12 @@ internal class DefaultCustomerSheetEventReporter @Inject constructor(
     private val analyticsRequestFactory: AnalyticsRequestFactory,
     @IOContext private val workContext: CoroutineContext,
 ) : CustomerSheetEventReporter {
-    override fun onInit(configuration: CustomerSheet.Configuration) {
+    override fun onInit(
+        configuration: CustomerSheet.Configuration,
+        integrationType: CustomerSheetIntegration.Type,
+    ) {
         fireEvent(
-            CustomerSheetEvent.Init(configuration)
+            CustomerSheetEvent.Init(configuration, integrationType)
         )
     }