+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/FormHelperTest.kt
@@ -128,6 +128,7 @@ internal class FormHelperTest {
                     value = 1099,
                     currencyCode = "usd",
                 ),
+                hasIntentToSetup = true,
                 billingDetails = PaymentSheet.BillingDetails(),
             )
         )